upload de foto completa, pendiente refactorizar
