+++ src/utils/firebase.ts
@@ -4,12 +4,12 @@ import { getFirestore } from 'firebase/firestore/lite';
 import { getStorage } from 'firebase/storage';
 
 const firebaseConfig = {
-  apiKey: 'AIzaSyCUW00Eg4gXDMdx4Fl0947uU9i9NeUgLkc',
-  authDomain: 'casa-nacional-judo.firebaseapp.com',
-  projectId: 'casa-nacional-judo',
-  storageBucket: 'casa-nacional-judo.appspot.com',
-  messagingSenderId: '356403942688',
-  appId: '1:356403942688:web:fb22fd197ab5b64356f345'
+    apiKey: 'AIzaSyCUW00Eg4gXDMdx4Fl0947uU9i9NeUgLkc',
+    authDomain: 'casa-nacional-judo.firebaseapp.com',
+    projectId: 'casa-nacional-judo',
+    storageBucket: 'casa-nacional-judo.appspot.com',
+    messagingSenderId: '356403942688',
+    appId: '1:356403942688:web:fb22fd197ab5b64356f345'
 };
 
 const app = initializeApp(firebaseConfig);
@@ -17,3 +17,4 @@ const auth = getAuth(app);
 const db = getFirestore(app);
 
 export { app, auth, db };
+export const storage = getStorage(app);
